LayoutTests/platform/mac/imported/w3c/web-platform-tests/uievents/keyboard/modifier-keys-expected.txt

Target

FAIL Check sending Shift key assert_equals: expected true but got false
FAIL Check sending Control key assert_equals: expected true but got false
FAIL Check sending Alt key assert_equals: expected true but got false
FAIL Check sending Meta key assert_equals: expected "Meta" but got "leftMeta"
